Load *only* external mods if the directory exists
If the "_externalMods" directory exists inside the mod folder, it will be used exclusively, and other mods in the load order file will be ignored.

This is enabled by default, and can be disabled in the ini by setting LoadExternalModsOnly to 0.

External mods can still be loaded with this option disabled, by putting them in the load order file in the desired priority (i.e. loading them as a normal mod).